+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- ChurchButton changed from using a Shape to SVG Image
+
 ### Fixed
 
 ### Removed